Updated status message for all script event classes, added alot of comments (still need more...).
Changed how window handles are obtained. (Every time a script event tries to use the window handle it searches for the newest window handle using the window title because the window handle changes quite often. This means that if the window title changes then the event will not work - there doesn't seem to be a 100% consistent window feature obtainable which would allow us to get the window handle every time, for now the window title is the closest thing I can think of.)

Updated main form UI a little bit.